pyproject.toml
[tool.poetry]
name = "meeting-pinger"
version = "0.1.0"
description = "Polls Google Calendar and pings via Slack DM before meetings"
authors = ["everstar"]
packages = [{include = "meeting_pinger"}]

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.100.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.2.0"
slack-bolt = "^1.18.0"
slack-sdk = "^3.26.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["poetry-core==1.9.1"]
build-backend = "poetry.core.masonry.api"
